Remove admin analytics blueprint and refactor analytics functionality into admin_bp
