+++ admin/__init__.py
@@ -3,4 +3,3 @@ from .admin_demo_bp import admin_demo_bp
 from .admin_org_bp import admin_org_bp
 from .admin_user_bp import admin_user_bp
 from .admin_recu_demo_bp import admin_recu_demo_bp
-from .admin_analytics_bp import admin_analytics_bp
