+++ admin/admin_bp.py
@@ -40,6 +40,35 @@ login_manager = LoginManager()
 login_manager.login_view = "auth.login"
 
 
+class DemoViewCount:
+    def __init__(self, demo_id, count):
+        self.id = demo_id
+        self.views = count
+
+    def __repr__(self):
+        return f"DemoViewCount({self.id}, {self.views})"
+
+    def __str__(self):
+        return f"Demo ID: {self.id}, Count: {self.views}"
+
+def count_per_demo(data):
+    demo_count = {}
+    for view in data:
+        demo_id = view.get("demo_id")
+        if demo_id in demo_count:
+            demo_count[demo_id] += 1
+        else:
+            demo_count[demo_id] = 1
+    
+    demo_count = [DemoViewCount(demo_id, count) for demo_id, count in demo_count.items()]
+    return demo_count
+
+
+from flask import Blueprint, render_template
+from utils.analytics import get_demo_views
+
+
+
 # User loader function
 @login_manager.user_loader
 def load_user(user_id):
@@ -99,6 +128,9 @@ def stats():
     # Count users by role
     role_counts = get_user_role_counts()
 
+    data = get_demo_views()
+    data = count_per_demo(data)
+    
     # Render statistics template
     return render_template(
         "admin/stats.html",
@@ -106,6 +138,7 @@ def stats():
         total_organizations=total_organizations,
         role_counts=role_counts,
         active_users=active_users,
+        data=data
     )
 
 
@@ -167,3 +200,12 @@ def manage_marquee():
         background_color=marquee_config["style"].split(": ")[1].split(";")[0],
         text_color=marquee_config["h2_style"].split(": ")[1],
     )
+
+
+
+@admin_bp.route('/admin/analytics')
+def admin_analytics():
+    data = get_demo_views()
+    data = count_per_demo(data)    
+    
+    return render_template('admin/analytics.html', data=data)
