+++ templates/admin/stats.html
@@ -3,6 +3,18 @@
 {% block styles %}
 <link rel="stylesheet" href="{{ url_for('static', filename='css/stats.css') }}" />
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" />
+<style>
+    .no-data {
+        color: #721c24;
+        font-weight: bold;
+        background: #f8d7da;
+        padding: 20px;
+        border: 1px dashed #f5c6cb;
+        border-radius: 5px;
+        text-align: center;
+        margin: 1rem 0;
+    }
+</style>
 {% endblock %}
 
 {% block main_content %}
@@ -42,10 +54,93 @@
         </div>
     </div>
 
+    <div class="analytics-container" style="margin-top: 40px;">
+        <h2 id="tabletitle">{{ _('Mielenosoitus analytiikka') }}</h2>
+
+        {% if data %}
+        <div class="table-container" style="background: none;">
+            <table>
+                <thead>
+                    <tr>
+                        <th>{{ _('Mielenosoituksen ID') }}</th>
+                        <th>{{ _('Katselukerrat') }}</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    {% for demo in data %}
+                    <tr>
+                        <td>{{ demo.id }}</td>
+                        <td>{{ demo.views }}</td>
+                    </tr>
+                    {% endfor %}
+                </tbody>
+            </table>
+        </div>
+        {% else %}
+        <div class="no-data">
+            <p>{{ _('Ei tietoja saatavilla.') }}</p>
+        </div>
+        {% endif %}
+    </div>
+
     <div style="margin-top: 20px;">
         <a href="{{ url_for('admin.admin_dashboard') }}" class="button">
             <i class="fas fa-arrow-left"></i>{{ _('Palaa hallintapaneeliin') }}
         </a>
     </div>
 </div>
+{% endblock %}
+{% block scripts %}
+<!-- DISABLE PREVIEW TEMPORARILY -->
+<!--
+<script>
+    document.addEventListener('DOMContentLoaded', function() {
+        const rows = document.querySelectorAll('tbody tr');
+
+        rows.forEach(row => {
+            row.addEventListener('mouseover', function() {
+                const demoId = this.querySelector('td').innerText;
+                const previewUrl = `/demonstration/${demoId}#main`;
+                
+                // Create a preview element
+                let preview = document.createElement('div');
+                preview.classList.add('preview');
+                preview.innerHTML = `<iframe src="${previewUrl}" frameborder="0"></iframe>`;
+                
+                // Append the preview to the body
+                document.body.appendChild(preview);
+                
+                // Position the preview near the row
+                const rect = this.getBoundingClientRect();
+                preview.style.top = `${rect.bottom + window.scrollY}px`;
+                preview.style.left = `${rect.left + window.scrollX}px`;
+            });
+
+            row.addEventListener('mouseout', function() {
+                // Remove the preview element
+                const preview = document.querySelector('.preview');
+                if (preview) {
+                    preview.remove();
+                }
+            });
+        });
+    });
+</script>
+
+<style>
+    .preview {
+        position: absolute;
+        width: 300px;
+        height: 200px;
+        border: 1px solid #ccc;
+        background: #fff;
+        box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+        z-index: 1000;
+    }
+
+    .preview iframe {
+        width: 100%;
+        height: 100%;
+    }
+</style> -->
 {% endblock %}
