+++ templates/sidebar.html
@@ -97,7 +97,7 @@
             </li>
             {% endif %}
             <li>
-                <a href="{{ url_for('admin_analytics.admin_analytics') }}">
+                <a href="{{ url_for('admin.admin_analytics') }}">
                     <span class="icon"><i class="fas fa-chart-bar"></i></span> Analytiikka
                 </a>
             </li>
@@ -141,7 +141,7 @@
     {% if current_user.has_permission("MANAGE_MARQUEE") %}
     <a href="{{ url_for('admin.manage_marquee') }}" class="hidden-icon"><i class="fas fa-scroll"></i></a> <!-- Marquee Management Hidden Icon -->
     {% endif %}
-    <a href="{{ url_for('admin_analytics.admin_analytics') }}" class="hidden-icon"><i class="fas fa-chart-bar"></i></a>
+    <a href="{{ url_for('admin.admin_analytics') }}" class="hidden-icon"><i class="fas fa-chart-bar"></i></a>
     <a href="{{ url_for('auth.logout') }}" class="hidden-icon"><i class="fas fa-sign-out-alt"></i></a>
 </div>
